+++ packages/jupyterlab-rjsf/src/typings.d.ts
@@ -1,14 +1,14 @@
-declare module '!!raw-loader!*.svg' {
+declare module '*.svg' {
   const script: string;
   export default script;
 }
 
-declare module 'react-jsonschema-form/lib/components/fields/ObjectField' {
+declare module '@rjsf/core/lib/components/fields/ObjectField' {
   import * as React from 'react';
   export default class ObjectField extends React.Component<any, any, any> {}
 }
 
-declare module 'react-jsonschema-form/lib/components/fields/StringField' {
+declare module '@rjsf/core/lib/components/fields/StringField' {
   import * as React from 'react';
   export default class StringField extends React.Component<any, any, any> {}
 }